12-voice-journal-ai/requirements.txt
anthropic>=0.40.0
openai-whisper>=20231117
